# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SIGAP API
  version: 1.0.0
  description: Geospatial infrastructure reporting platform API
servers:
  - url: https://sigap.live/api
    description: Production
paths:
...
  /export/geojson:
    get:
      operationId: exportReportsGeoJSON
      tags: [export]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: category_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: GeoJSON FeatureCollection
          content:
            application/geo+json:
              schema:
                $ref: "#/components/schemas/GeoJSONFeatureCollection"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /export/csv:
    get:
      operationId: exportReportsCSV
      tags: [export]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: category_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
                format: binary
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==================== PUBLIC ====================
...
  /audit/search:
    get:
      operationId: searchAuditLog
      tags: [audit]
      security:
        - bearerAuth: []
      parameters:
        - name: actor_id
          in: query
          schema:
            type: string
            format: uuid
        - name: action
          in: query
          schema:
            type: string
        - name: object_type
          in: query
          schema:
            type: string
        - name: object_id
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        actor:
                          type: string
                          format: uuid
                        actor_role:
                          type: string
                        action:
                          type: string
                        object_type:
                          type: string
                        object_id:
                          type: string
                        before:
                          type: object
                          nullable: true
                        after:
                          type: object
                          nullable: true
                        reason:
                          type: string
                          nullable: true
                        created_at:
                          type: string
                          format: date-time
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /audit/export:
    get:
      operationId: exportAuditLog
      tags: [audit]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Audit log export
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==================== OUTBOX ====================
...
  /auditor/audit-search:
    get:
      operationId: auditorSearchAudit
      tags: [auditor]
      security:
        - bearerAuth: []
      parameters:
        - name: actor_id
          in: query
          schema:
            type: string
            format: uuid
        - name: action
          in: query
          schema:
            type: string
        - name: object_type
          in: query
          schema:
            type: string
        - name: object_id
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        "200":
          description: Audit entries
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ReportStatus:
      type: string
      enum:
        - draft
        - submitted
        - under_review
        - verified
        - assigned
        - in_progress
        - resolved
        - closed
        - rejected
        - duplicate_merged
        - needs_survey
        - needs_completion
        - out_of_scope
        - merged
        - separated
        - escalated
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
        details:
          type: object
          nullable: true

    # Auth
...
    GeoJSONFeatureCollection:
      type: object
      required: [type, features]
      properties:
        type:
          type: string
          const: FeatureCollection
        features:
          type: array
          items:
            $ref: "#/components/schemas/GeoJSONFeature"
    GeoJSONFeature:
      type: object
      required: [type, geometry, properties]
      properties:
        type:
          type: string
          const: Feature
        geometry:
          type: object
          oneOf:
            - $ref: "#/components/schemas/GeoJSONPoint"
            - $ref: "#/components/schemas/GeoJSONPolygon"
        properties:
          type: object
          required: [id, status, category_id, description, created_at]
          properties:
            id:
              type: string
              format: uuid
            status:
              $ref: "#/components/schemas/ReportStatus"
            category_id:
              type: string
              format: uuid
            description:
              type: string
            severity:
              type: integer
              nullable: true
            created_at:
              type: string
              format: date-time
    GeoJSONPoint:
      type: object
      required: [type, coordinates]
      properties:
        type:
          type: string
          const: Point
        coordinates:
          type: array
          items:
            type: number
          minItems: 2
          maxItems: 2
    GeoJSONPolygon:
      type: object
      required: [type, coordinates]
      properties:
        type:
          type: string
          const: Polygon
        coordinates:
          type: array
          items:
            type: array
            items:
              type: array
              items:
                type: number
              minItems: 2